api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.35.0
python-multipart==0.0.6
supabase==2.0.2
python-dotenv==1.0.0
